requirements.txt
pyTelegramBotAPI
telethon
python-dotenv
openai
